fix: correct 2fa page detection
Ran into an issue due to slight differences between versions and features
of Neptun at different educational institutes.
The previous detection only worked if the 2fa tab was the first in the
settings list at the page load.
We don't really need the tab to be visible in order to modify it,
so the active tab restriction is removed.